requirements.txt
anthropic>=0.25.0
PyMuPDF>=1.24.0
openpyxl>=3.1.0
Pillow>=10.0.0
click>=8.1.0
python-dotenv>=1.0.0
# supabase and its deps — install with:
#   pip install supabase --no-deps
#   pip install postgrest gotrue realtime supafunc storage3==0.8.1 --no-deps
#   pip install deprecation yarl "websockets>=11,<16" PyJWT python-dateutil six httpx
supabase>=2.0.0
storage3==0.8.1
PyJWT>=2.0.0
six
python-dateutil
deprecation
yarl>=1.22.0
